Refractor CNNBlock
