+++ models/resnet.py
@@ -1,4 +1,5 @@
 import torch
+from sublayers import CNNBlock
 
 class ResNet6(torch.nn.Module):
    def __init__(self):
@@ -45,14 +46,3 @@ class ResNetBlock(torch.nn.Module):
          X = X[:, :, ::self.stride, ::self.stride] # (B, C, H, W)
       return Z + X
    
-
-class CNNBlock(torch.nn.Module):
-   def __init__(self, in_channels, out_channels, kernel_size, stride, padding):
-      super().__init__()
-      self.layers = torch.nn.Sequential(
-         torch.nn.Conv2d(in_channels, out_channels, kernel_size, stride, padding),
-         torch.nn.BatchNorm2d(out_channels),
-         torch.nn.PReLU(),
-      )
-   def forward(self, X):
-      return self.layers(X)
